requirements.txt
azure-iot-device==2.14.0
certifi==2024.7.4
charset-normalizer==3.3.2
deprecation==2.1.0
idna==3.8
janus==1.0.0
packaging==24.1
paho-mqtt==1.6.1
PySocks==1.7.1
requests==2.32.3
requests-unixsocket2==0.4.1
scapy==2.5.0
typing_extensions==4.12.2
urllib3==2.2.2
